extracted EventCollectionRunner
