+++ boudicca.base/eventcollector-client/src/main/kotlin/base/boudicca/api/eventcollector/EventCollectorScheduler.kt
@@ -50,7 +50,7 @@ class EventCollectorScheduler(
     }
 
     fun runOnce() {
-        getEventCollectorCoordinator().runOnce()
+        getEventCollectorCoordinator().getEventCollectionRunner().run()
     }
 
     fun getCollectors(): List<EventCollector> {
